fix compilation against Qt 6.6.1 for macos

+++ components/contentselector/model/esmfile.cpp
@@ -113,7 +113,7 @@ void ContentSelectorModel::EsmFile::setFileProperty(const FileProperty prop, con
             break;
 
         case FileProperty_Format:
-            mVersion = value.toInt();
+            mVersion = value;
             break;
 
         case FileProperty_DateModified:
